requirements.txt
openai>=0.27.0
python-dotenv>=0.19.0
sqlite3
pandas>=1.3.0
numpy>=1.21.0
networkx
openpyxl
openai>=0.27.0
python-dotenv>=0.19.0
